Debounce guesses
Previously, a user typing "and" into a poem that contained the word "a" would have the "a" entered straight away, and the guess box cleared. This would then result in them having "nd" remaining in the guess box and being a bit confused.

Set a 350ms debounce so this doesn't happen unless the user stops typing. The exception is where the user presses space, which causes the guess to update straight away.

squash

squash